antilles-core/openHPC_web_project/requirements.txt
Django~=1.11.0
djangorestframework~=3.7
pamela>=0.3
cryptography>=2.0
six>=1.10
setuptools>=36.0
PyJWT>=1.5.2
celery~=4.1
django-celery-results~=1.0.1
requests>=2.6
pika>=0.11
PyYAML>=3.10
PasteDeploy>=1.5
Paste>=2.0
pandas>=0.19
WeasyPrint~=0.42.3
xlwt>=1.2
jsonschema>=2.6.0
influxdb>=5.0.0
psutil>=5.4
